requirements.txt
apex==0.1
Deprecated==1.2.10
fvcore==0.1.1.post200513
imageio-ffmpeg==0.4.1
imgaug==0.4.0
numpy==1.17.5
pandas==0.25.1
Pillow==7.1.2
pycocotools==2.0
pypng==0.0.20
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.1.0
tensorboard==1.15.0
torch==1.5.1
torchsummary==1.5.1
torchvision==0.6.0a0+35d732a
tornado==6.0.3
tqdm==4.43.0
yacs==0.1.6
